requirements.txt
backtrader==1.9.74.123
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
DateTime==4.3
dnspython==1.16.0
eventlet==0.25.0
Flask==1.1.1
Flask-Cors==3.0.8
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
monotonic==1.5
numpy==1.17.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
zope.interface==4.6.0
